move `struct_wf` to `wf_goals`
